requirements.txt
mysql_connector_repackaged==0.3.1
streamlit==1.45.0
